yolo asci/requirements.txt
fastapi
uvicorn
pillow
ultralytics
python-multipart
easyocr
opencv-python-headless
kafka-python